added tests to rack field

+++ src/Web.Api/VueClient/src/components/RackForm.vue
@@ -11,13 +11,12 @@
                               item-value=""
                               :return-object="false"
                               label="Datacenter"
-                              placeholder="Select a datacenter or all datacenters"
-                              >
+                              placeholder="Select a datacenter or all datacenters">
                     </v-select>
-                    <v-text-field v-model="range.start" 
+                    <v-text-field v-model="range.start"
                                   label="Start"
                                   placeholder="Enter the start of a range of racks (i.e. A1)"></v-text-field>
-                    <v-text-field v-model="range.end" 
+                    <v-text-field v-model="range.end"
                                   label="End"
                                   placeholder="Enter the end of a range of racks (i.e. Z20)"></v-text-field>
 
@@ -27,9 +26,22 @@
                         <v-btn v-if="admin" class="mr-4" color="primary" @click="createInRange">Add Racks</v-btn>
                         <v-btn v-if="admin" class="mr-4" color="primary" @click="deleteInRange">Delete Racks</v-btn>
                     </v-row>
-                    
+
                 </form>
             </v-card-text>
+            <v-snackbar v-model="updateSnackbar.show"
+                        :bottom=true
+                        class="black--text"
+                        :color="updateSnackbar.color"
+                        :timeout=5000>
+                {{updateSnackbar.message}}
+                <v-btn dark
+                       class="black--text"
+                       text
+                       @click="updateSnackbar.show = false">
+                    Close
+                </v-btn>
+            </v-snackbar>
         </v-card>
     </v-container>
 </template>
@@ -42,12 +54,20 @@ export default {
     inject: ['rackRepository', 'datacenterRepository'],
     data: () => ({
         loading: true,
+        updateSnackbar: {
+                    show: false,
+                    message: '',
+                    color: ''
+                },
         range: {
             start: '',
             end: ''
         },
         datacenters: [],
         selectedDatacenter: '',
+        rules: {
+            rackRules: v => /^[a-zA-Z][0-9]+$/.test(v),
+        },
     }),
     computed: {
         admin() {
@@ -66,9 +86,15 @@ export default {
             this.loading = false;
          },
         viewDiagram() {
+            if (this.validation() != 0) {
+                return;
+            }
             this.$router.push({ name: "RackDiagram", query: { start: this.range.start, end: this.range.end, datacenter: this.selectedDatacenter } });
         },
         async createInRange() {
+            if (this.validation() != 0) {
+                return;
+            }
             var searchDatacenter = this.datacenters.find(o => o.description === this.selectedDatacenter);
 
             await this.rackRepository.createInRange(this.range.start, this.range.end, searchDatacenter.id)
@@ -85,6 +111,9 @@ export default {
             //this.$emit('update-create'); // trigger event to update rack-table
         },
         async deleteInRange() {
+            if (this.validation() != 0) {
+                return;
+            }
             var searchDatacenter = this.datacenters.find(o => o.description === this.selectedDatacenter);
 
             await this.rackRepository.deleteInRange(this.range.start, this.range.end, searchDatacenter.id)
@@ -99,6 +128,17 @@ export default {
 
             // This might slow things down if we have a lot of racks to get from the backend !!!
             //this.$emit('update-delete'); // trigger event to update rack-table
+        },
+        validation() {
+            var count = 0;
+            this.updateSnackbar.message = '';
+            if (!(/^[a-zA-Z][0-9]+$/.test(this.range.start)) || !(/^[a-zA-Z][0-9]+$/.test(this.range.start))) {
+                this.updateSnackbar.show = true;
+                this.updateSnackbar.color = 'red lighten-4';
+                this.updateSnackbar.message = this.updateSnackbar.message + 'Rack action failed. Rack name must be a letter followed by a number ';
+                count++;
+            }
+            return count;
         }
     }
 }
